Move validate tests to their own file
